refactor(catalog): isolate purchase workflow
