--- Emulator/src/main/java/com/eu/habbo/habbohotel/catalog/CatalogPurchaseCommand.java
@@ -1,3 +1,0 @@
-package com.eu.habbo.habbohotel.catalog;
-
-public record CatalogPurchaseCommand(int pageId, int itemId, String extraData, int count) {}
