+++ Emulator/src/main/java/com/eu/habbo/messages/incoming/catalog/CatalogPurchaseCommand.java
@@ -1,0 +1,3 @@
+package com.eu.habbo.messages.incoming.catalog;
+
+record CatalogPurchaseCommand(int pageId, int itemId, String extraData, int count) {}
