+++ Emulator/src/main/java/com/eu/habbo/messages/incoming/catalog/CatalogPurchaseCommandReader.java
@@ -1,6 +1,5 @@
 package com.eu.habbo.messages.incoming.catalog;
 
-import com.eu.habbo.habbohotel.catalog.CatalogPurchaseCommand;
 import com.eu.habbo.messages.ClientMessage;
 
 final class CatalogPurchaseCommandReader {
