+++ Emulator/src/test/java/com/eu/habbo/messages/incoming/catalog/CatalogPurchasePageResolverTest.java
@@ -1,4 +1,4 @@
-package com.eu.habbo.habbohotel.catalog;
+package com.eu.habbo.messages.incoming.catalog;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNull;
@@ -7,17 +7,20 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import com.eu.habbo.habbohotel.catalog.CatalogItem;
+import com.eu.habbo.habbohotel.catalog.CatalogPage;
+import com.eu.habbo.habbohotel.catalog.CatalogPageLayouts;
 import java.util.concurrent.atomic.AtomicBoolean;
 import org.junit.jupiter.api.Test;
 
-class CatalogPurchasePageServiceTest {
+class CatalogPurchasePageResolverTest {
 
     @Test
     void unknownSearchItemDoesNotEvaluatePageAccess() {
         AtomicBoolean accessEvaluated = new AtomicBoolean();
-        CatalogPurchasePageService service = service(null, null);
+        CatalogPurchasePageResolver resolver = resolver(null, null);
 
-        CatalogPage resolved = service.resolve(new CatalogPurchaseCommand(-1, 404, "", 1), page -> {
+        CatalogPage resolved = resolver.resolve(new CatalogPurchaseCommand(-1, 404, "", 1), page -> {
             accessEvaluated.set(true);
             return true;
         });
@@ -32,13 +35,13 @@ class CatalogPurchasePageServiceTest {
         CatalogPage page = mock(CatalogPage.class);
         when(item.getOfferId()).thenReturn(72);
         when(item.getPageId()).thenReturn(19);
-        CatalogPurchasePageService.CatalogLookup lookup = mock(CatalogPurchasePageService.CatalogLookup.class);
+        CatalogPurchasePageResolver.CatalogLookup lookup = mock(CatalogPurchasePageResolver.CatalogLookup.class);
         when(lookup.findItem(72)).thenReturn(item);
         when(lookup.findPage(19)).thenReturn(page);
-        CatalogPurchasePageService service = new CatalogPurchasePageService(lookup);
+        CatalogPurchasePageResolver resolver = new CatalogPurchasePageResolver(lookup);
 
         CatalogPage resolved =
-                service.resolve(new CatalogPurchaseCommand(-1, 72, "", 1), candidate -> candidate == page);
+                resolver.resolve(new CatalogPurchaseCommand(-1, 72, "", 1), candidate -> candidate == page);
 
         assertSame(page, resolved);
         verify(lookup).findPage(19);
@@ -47,11 +50,11 @@ class CatalogPurchasePageServiceTest {
     @Test
     void directPageUsesThePageIdWithoutAnItemLookup() {
         CatalogPage page = mock(CatalogPage.class);
-        CatalogPurchasePageService.CatalogLookup lookup = mock(CatalogPurchasePageService.CatalogLookup.class);
+        CatalogPurchasePageResolver.CatalogLookup lookup = mock(CatalogPurchasePageResolver.CatalogLookup.class);
         when(lookup.findPage(11)).thenReturn(page);
-        CatalogPurchasePageService service = new CatalogPurchasePageService(lookup);
+        CatalogPurchasePageResolver resolver = new CatalogPurchasePageResolver(lookup);
 
-        assertSame(page, service.resolve(new CatalogPurchaseCommand(11, 72, "", 1), candidate -> true));
+        assertSame(page, resolver.resolve(new CatalogPurchaseCommand(11, 72, "", 1), candidate -> true));
     }
 
     @Test
@@ -59,17 +62,17 @@ class CatalogPurchasePageServiceTest {
         CatalogPage page = mock(CatalogPage.class);
         when(page.getLayout()).thenReturn(CatalogPageLayouts.club_gift.name());
         AtomicBoolean accessEvaluated = new AtomicBoolean();
-        CatalogPurchasePageService service = service(null, page);
+        CatalogPurchasePageResolver resolver = resolver(null, page);
 
-        assertNull(service.resolve(new CatalogPurchaseCommand(11, 72, "", 1), candidate -> {
+        assertNull(resolver.resolve(new CatalogPurchaseCommand(11, 72, "", 1), candidate -> {
             accessEvaluated.set(true);
             return true;
         }));
         assertFalse(accessEvaluated.get());
     }
 
-    private static CatalogPurchasePageService service(CatalogItem item, CatalogPage page) {
-        return new CatalogPurchasePageService(new CatalogPurchasePageService.CatalogLookup() {
+    private static CatalogPurchasePageResolver resolver(CatalogItem item, CatalogPage page) {
+        return new CatalogPurchasePageResolver(new CatalogPurchasePageResolver.CatalogLookup() {
             @Override
             public CatalogItem findItem(int itemId) {
                 return item;
